requirements.txt
streamlit>=1.31,<2
numpy>=1.26,<3
scipy>=1.11,<2
plotly>=5.18,<6
pillow>=10,<12
